requirements.txt
fastapi
uvicorn
Pillow
psycopg2-binary
redis
